api: consolidate tests
* move tests to follow the pattern <filename>_test.go
* dedup some duplicated tests
* replace doGET with get() and newAPIServer with server()
